+++ packages/ricos-content/web/lib/converters.ts
@@ -5,7 +5,6 @@ export {
   ensureRicosContent,
   TO_RICOS_NODE_TYPE,
 } from '../src/converters/draft';
-export { TO_DRAFT_PLUGIN_TYPE_MAP, TO_RICOS_PLUGIN_TYPE_MAP } from '../src/consts';
 export { fromPlainText, toPlainText } from '../src/converters/plainText';
 export * from '../src/converters/nodeUtils';
 export * from '../src/converters/jsonUtils';
