+++ packages/ricos-content/web/src/converters/draft/consts.ts
@@ -23,6 +23,49 @@ import {
   EMBED_TYPE,
   RICOS_NODE_TYPE_TO_DATA_FIELD,
   EXTERNAL,
+  RICOS_DIVIDER_TYPE,
+  RICOS_LINK_TYPE,
+  RICOS_TEXT_COLOR_TYPE,
+  RICOS_TEXT_HIGHLIGHT_TYPE,
+  RICOS_FILE_TYPE,
+  RICOS_GALLERY_TYPE,
+  RICOS_GIPHY_TYPE,
+  RICOS_HASHTAG_TYPE,
+  RICOS_HTML_TYPE,
+  RICOS_CODE_BLOCK_TYPE,
+  RICOS_IMAGE_TYPE,
+  RICOS_VIDEO_TYPE,
+  RICOS_EMOJI_TYPE,
+  VERTICAL_EMBED_TYPE,
+  TEXT_COLOR_TYPE,
+  TEXT_HIGHLIGHT_TYPE,
+  SOUND_CLOUD_TYPE,
+  RICOS_SOUND_CLOUD_TYPE,
+  RICOS_UNDO_REDO_TYPE,
+  RICOS_SPOILER_TYPE,
+  RICOS_INDENT_TYPE,
+  RICOS_LINE_SPACING_TYPE,
+  RICOS_POLL_TYPE,
+  RICOS_MENTION_TYPE,
+  RICOS_COLLAPSIBLE_LIST_TYPE,
+  EMOJI_TYPE,
+  INDENT_TYPE,
+  HEADINGS_DROPDOWN_TYPE,
+  RICOS_MAP_TYPE,
+  RICOS_VERTICAL_EMBED_TYPE,
+  EXTERNAL_LINK_TYPE,
+  UNDO_REDO_TYPE,
+  HASHTAG_TYPE,
+  LINE_SPACING_TYPE,
+  CODE_BLOCK_TYPE,
+  HEADERS_MARKDOWN_TYPE,
+  RICOS_HEADINGS_DROPDOWN_TYPE,
+  RICOS_LINK_BUTTON_TYPE,
+  RICOS_LINK_PREVIEW_TYPE,
+  RICOS_EXTERNAL_LINK_TYPE,
+  RICOS_TABLE_TYPE,
+  RICOS_ACTION_BUTTON_TYPE,
+  RICOS_HEADERS_MARKDOWN_TYPE,
 } from '../../consts';
 import { Decoration_Type, Node_Type } from 'ricos-schema';
 import { fromEntries } from '../../utils';
@@ -183,6 +226,67 @@ export type DraftGalleryStyles = {
   galleryThumbnailsAlignment?: string;
   imageMargin?: number;
   thumbnailSpacings?: number;
+};
+
+export const TO_DRAFT_PLUGIN_TYPE_MAP = {
+  [RICOS_DIVIDER_TYPE]: DIVIDER_TYPE,
+  [RICOS_FILE_TYPE]: FILE_UPLOAD_TYPE,
+  [RICOS_GALLERY_TYPE]: GALLERY_TYPE,
+  [RICOS_GIPHY_TYPE]: GIPHY_TYPE,
+  [RICOS_HTML_TYPE]: HTML_TYPE,
+  [RICOS_IMAGE_TYPE]: IMAGE_TYPE,
+  [RICOS_VIDEO_TYPE]: VIDEO_TYPE,
+  [RICOS_POLL_TYPE]: POLL_TYPE,
+  [RICOS_LINK_TYPE]: LINK_TYPE,
+  [RICOS_MENTION_TYPE]: MENTION_TYPE,
+  [RICOS_TEXT_HIGHLIGHT_TYPE]: TEXT_HIGHLIGHT_TYPE,
+  [RICOS_TEXT_COLOR_TYPE]: TEXT_COLOR_TYPE,
+  [DIVIDER_TYPE]: DIVIDER_TYPE,
+  [FILE_UPLOAD_TYPE]: FILE_UPLOAD_TYPE,
+  [GALLERY_TYPE]: GALLERY_TYPE,
+  [GIPHY_TYPE]: GIPHY_TYPE,
+  [HTML_TYPE]: HTML_TYPE,
+  [IMAGE_TYPE]: IMAGE_TYPE,
+  [VIDEO_TYPE]: VIDEO_TYPE,
+  [POLL_TYPE]: POLL_TYPE,
+  [TEXT_HIGHLIGHT_TYPE]: TEXT_HIGHLIGHT_TYPE,
+  [TEXT_COLOR_TYPE]: TEXT_COLOR_TYPE,
+  [RICOS_INDENT_TYPE]: RICOS_INDENT_TYPE,
+  [RICOS_LINE_SPACING_TYPE]: RICOS_LINE_SPACING_TYPE,
+  [EXTERNAL]: EXTERNAL,
+};
+
+export const TO_RICOS_PLUGIN_TYPE_MAP = {
+  [DIVIDER_TYPE]: RICOS_DIVIDER_TYPE,
+  [FILE_UPLOAD_TYPE]: RICOS_FILE_TYPE,
+  [GALLERY_TYPE]: RICOS_GALLERY_TYPE,
+  [GIPHY_TYPE]: RICOS_GIPHY_TYPE,
+  [HTML_TYPE]: RICOS_HTML_TYPE,
+  [IMAGE_TYPE]: RICOS_IMAGE_TYPE,
+  [VIDEO_TYPE]: RICOS_VIDEO_TYPE,
+  [POLL_TYPE]: RICOS_POLL_TYPE,
+  [LINK_TYPE]: RICOS_LINK_TYPE,
+  [MENTION_TYPE]: RICOS_MENTION_TYPE,
+  [COLLAPSIBLE_LIST_TYPE]: RICOS_COLLAPSIBLE_LIST_TYPE,
+  [ACTION_BUTTON_TYPE]: RICOS_ACTION_BUTTON_TYPE,
+  [LINK_BUTTON_TYPE]: RICOS_LINK_BUTTON_TYPE,
+  [CODE_BLOCK_TYPE]: RICOS_CODE_BLOCK_TYPE,
+  [EMOJI_TYPE]: RICOS_EMOJI_TYPE,
+  [HASHTAG_TYPE]: RICOS_HASHTAG_TYPE,
+  [HEADERS_MARKDOWN_TYPE]: RICOS_HEADERS_MARKDOWN_TYPE,
+  [INDENT_TYPE]: RICOS_INDENT_TYPE,
+  [LINE_SPACING_TYPE]: RICOS_LINE_SPACING_TYPE,
+  [TABLE_TYPE]: RICOS_TABLE_TYPE,
+  [EXTERNAL_LINK_TYPE]: RICOS_EXTERNAL_LINK_TYPE,
+  [LINK_PREVIEW_TYPE]: RICOS_LINK_PREVIEW_TYPE,
+  [SPOILER_TYPE]: RICOS_SPOILER_TYPE,
+  [UNDO_REDO_TYPE]: RICOS_UNDO_REDO_TYPE,
+  [HEADINGS_DROPDOWN_TYPE]: RICOS_HEADINGS_DROPDOWN_TYPE,
+  [MAP_TYPE]: RICOS_MAP_TYPE,
+  [SOUND_CLOUD_TYPE]: RICOS_SOUND_CLOUD_TYPE,
+  [TEXT_COLOR_TYPE]: RICOS_TEXT_COLOR_TYPE,
+  [TEXT_HIGHLIGHT_TYPE]: RICOS_TEXT_HIGHLIGHT_TYPE,
+  [VERTICAL_EMBED_TYPE]: RICOS_VERTICAL_EMBED_TYPE,
 };
 
 export { RICOS_NODE_TYPE_TO_DATA_FIELD }; // exists here to avoid breaking API
